2017 (#143)
* radical stylectomy

* integrated bootstrap

IT WAS A BIG YAK

* strip it all out

* restructuring by degrees

moved most partials into *_old folder; bringing over as I edit them

* moving the pages into a collection

* starting cleanup/stripping of html files

* svgs update

* hiding posts to avoid compiling issues

* starting restructuring

stripping out old classes, modifying html/includes structure

* change to gem-based xml feed (#133)

so I don't have to write a feed page anymore

* straightening up

* forgot to clear these

* basic top nav

* getting the three-bar icon to show up

edited to remove "str-replace" function

* forgot to remove the `!default` marker

* basic structure front page (#135)

* home page structure

copy/paste from bootstrap website, starting to liquid-fy it

* made double-looping front page

3 columns per row, 3 rows per page = 9 cards on the front page :)

* different selection of posts to for testing

* added to get jekyll to compile

* home page central layout is more-or-less done

still need to do footer and general styling of colors etc

* footer reorganize and style

* colors, fonts and spacing on home page (#136)

* Posts structure (#139)

* started restructure of post

* picking different posts as examples

* toc, asides, footnotes done(ish) Images next!

* warning warning (#137)

* more example posts

* medium image structure

BIG rearrangement to restructure the medium-sized images. Had to edit
the posts, but hopefully won't have to do it again. Until I get to the
small images I guess.

* fixed site headline

* minor edits to typography

* Small images restructure (#138)

* working on srcset and getting all discombobulated

* srcset picture maker is made and installed :)

* a few more notes

* cleaned up some layout (still needs work)

* started building figures include

* DID IT. I can now use an include in my markdown for figure srcset markup

* tweak variable name

* image size kludges

now I'm out of my element and just trying to get anything to look close
to what I want

* log

* I MADE A WORKPLAN

* AND I'VE ALREADY CHANGED IT

* housekeeping (#140)

* one last note

* Post image (#142)

* dailylog

* post lightGallery update

# Conflicts:
#	notes to myself.txt

* update for today

* decided more-or-less on `ekko-lightbox`

some styling tweaks still need to be made, but it's close to ready

* new image naming structure

* today's notes

* notes and starting on updating the posting guide

* Lightbox (#141)

* housekeeping

* install ekko-lightbox

* rough test of `ekko-lightbox` as an include

It's kludgy, but seems to mostly work. Styles need a *lot* of work yet.

* styling of modal for lightbox

* moving posts around

Editing for the new figure method. Some have been pushed to `drafts`
because they're outdated (how I used to do images). One related to a
gallery is waiting until I redo the gallery stuff.

* tweaks as I find issues in the image related stuff

* updated posts

all done except for gallery-related ones

* tweaks, mostly

better handling of video includes, minor styling, place for the
orphaned posts and images that won't break jekyll

* housekeeping

* Clean up posts styling

* always more notes

* working away on the standalone pages

there's a lot of junk to get through...

* portfolio gallery extraction

made the index of portfolio galleries call a new `port-index` template
to make it easier to separate out the styling and logic

* restructuring gallery organization

* add sorting method to allow modified dates to supersede post dates

loops over all documents (posts and collection items) and assigns the
later of the mod date or orig date (from the YAML header of the item).
Used on `home` to bring galleries into the date-based flow on the front
page.

* working through the gallery "posts" and the posts about galleries

Still not sure on the location of the gallery images, but the basic
restructure is ok (all info in the YAML header of the gallery "post")

* new gallery page structure

This removes the data file of galleries, and the logic used to cycle
through them to get an index of all the galleries. Using more 'native'
jekyll code to do it now.

* moving orphaned files prior to deleting them

Because images aren't tracked in this repo

* removed mod dates

was using them for tests

* new version to do subdirectories

* update to new gallery/og-image structures

* title change

* more notes

* current notes before branch to paginate

* pagination is working

there are ugly bits, but I'm willing to deal with it right now

* notes updated

* ready for prime time!

There are a few bits still hanging (search, instagram, square) but
those were issues before now.

* remove some duplication

* I can't stop messing with it

* removed the unused items